feat(HEK#17): mueve icons a su propio archivo
Actualiza imports
