+++ src/icons/Icons.tsx
@@ -63,45 +63,3 @@ export const IGIcon = () => (
     <use href="/icons.svg#icon-instagram"></use>
   </svg>
 )
-
-export const GridIcon = () => (
-  <svg fill="none" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-view-grid"></use>
-  </svg>
-);
-
-export const ListIcon = () => (
-  <svg fill="none" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-view-list"></use>
-  </svg>
-);
-
-export const GridIconFilled = () => (
-  <svg fill="none" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-view-grid-filled"></use>
-  </svg>
-);
-
-export const ListIconFilled = () => (
-  <svg fill="none" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-view-list-filled"></use>
-  </svg>
-);
-
-export const StarIcon = () => (
-  <svg fill="none" height="16" viewBox="0 0 16 16" width="16" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-star-active"></use>
-  </svg>
-);
-
-export const StarEmptyIcon = () => (
-  <svg fill="none" height="16" viewBox="0 0 16 16" width="16" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#icon-star"></use>
-  </svg>
-);
-
-export const ChevronDownIcon = () => (
-  <svg fill="none" height="16" viewBox="0 0 16 16" width="16" xmlns="http://www.w3.org/2000/svg">
-    <use href="/icons.svg#chevron-down"></use>
-  </svg>
-);
